vue demo

+++ other/http/this.sql
@@ -1,7 +1,7 @@
 -- user_info
 create table user_info(
     id int(6) not null auto_increment,
-    uid int(10) not null,
+    uid BIGINT not null,
     name char(255) not null,
     profile_image_url varchar(1024) not null,
     verified int(1) not null,
@@ -10,7 +10,7 @@ create table user_info(
 
 create table photos(
     id int not null auto_increment,
-    uid int(10) not null,
+    uid BIGINT not null,
     photo_id char(20) not null,
     album_id char(20) not null,
     pid char(32) not null,
